feat: #526 organize Project API
Note that, in addition to organizing the exports and imports, we move PkgJson to
schemas.

The idea of this commit is to leave types where they belong (f.e: ProjectType
inside Project) and then export all from the manifest index.ts
