+++ packages/liferay-js-toolkit-core/src/project/VersionInfo.ts
@@ -1,0 +1,12 @@
+/**
+ * SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ */
+
+/** Plugin version information */
+export interface VersionInfo {
+	version: string;
+	path: string;
+}
+
+export default VersionInfo;
