+++ packages/liferay-js-toolkit-core/src/project/__tests__/Rules.test.ts
@@ -4,7 +4,7 @@
  */
 
 import FilePath from '../../file-path';
-import Rules from '../rules';
+import Rules from '../Rules';
 
 describe('single rule', () => {
 	it('test alone', () => {
